PHP 8.1 - cast in _RC4 before sending to str_repeat

+++ include/tcpdf_static.php
@@ -506,7 +506,7 @@ class TCPDF_STATIC {
 			return $out;
 		}
 		if ($last_enc_key != $key) {
-			$k = str_repeat($key, ((256 / strlen($key)) + 1));
+			$k = str_repeat($key, (int) ((256 / strlen($key)) + 1));
 			$rc4 = range(0, 255);
 			$j = 0;
 			for ($i = 0; $i < 256; ++$i) {
